refactor: upgrade versions and break down components into smaller sub-components in frontend
