+++ frontend/src/components/common/index.js
@@ -1,0 +1,1 @@
+export { default as AlertComponent } from './AlertComponent'
