+++ frontend/src/components/forms/index.js
@@ -1,0 +1,1 @@
+export { default as PasswordField } from './PasswordField'
