+++ frontend/src/components/navigation/NavigationDrawer.js
@@ -1,0 +1,54 @@
+import React from 'react'
+import { Box, Divider, Drawer, List, ListItem, ListItemButton, ListItemText, Typography } from '@mui/material'
+
+const drawerWidth = 240
+
+const navItems = ["Home", "About"]
+
+const routerMap = {
+  "Home": "/",
+  "About": "/about",
+}
+
+const NavigationDrawer = ({ mobileOpen, handleDrawerToggle, navigate, window }) => {
+  const drawer = (
+    <Box onClick={handleDrawerToggle} sx={{ textAlign: 'center' }}>
+      <Typography variant="h6" sx={{ my: 2 }}>
+        Client List
+      </Typography>
+      <Divider />
+      <List>
+        {navItems.map((item) => (
+          <ListItem key={item} disablePadding>
+            <ListItemButton sx={{ textAlign: 'center' }} onClick={() => navigate(routerMap[item])}>
+              <ListItemText primary={item} />
+            </ListItemButton>
+          </ListItem>
+        ))}
+      </List>
+    </Box>
+  )
+
+  const container = window !== undefined ? () => window().document.body : undefined
+
+  return (
+    <Box component="nav">
+      <Drawer
+        container={container}
+        variant="temporary"
+        open={mobileOpen}
+        onClose={handleDrawerToggle}
+        ModalProps={{
+          keepMounted: true, // Better open performance on mobile.
+        }}
+        sx={{
+          display: { xs: 'block', sm: 'none' },
+          '& .MuiDrawer-paper': { boxSizing: 'border-box', width: drawerWidth },
+        }}>
+        {drawer}
+      </Drawer>
+    </Box>
+  )
+}
+
+export default NavigationDrawer
