+++ frontend/src/components/navigation/index.js
@@ -1,0 +1,2 @@
+export { default as NavigationBar } from './NavigationBar'
+export { default as NavigationDrawer } from './NavigationDrawer'
