+++ frontend/src/components/table/index.js
@@ -1,0 +1,8 @@
+export { default as DataTableHead } from './DataTableHead'
+export { default as DataTableRowDisplay } from './DataTableRowDisplay'
+export { default as DataTableRowInEditMode } from './DataTableRowInEditMode'
+export { default as DataTableRowInAddMode } from './DataTableRowInAddMode'
+export { default as ButtonCellEdit } from './ButtonCellEdit'
+export { default as ButtonCellDisabled } from './ButtonCellDisabled'
+export { default as InputInEditMode } from './InputInEditMode'
+export { default as PaginationFooter } from './PaginationFooter'
