+++ pom.xml
@@ -14,7 +14,7 @@
     <name>springboot-react-demo</name>
     <description>springboot-react-demo</description>
     <properties>
-        <java.version>11</java.version>
+        <java.version>17</java.version>
     </properties>
     <dependencies>
         <dependency>
